working on bank conflicts in the 'bitonic merge sort' algorithm

+++ Nutty/base/reduce.h
@@ -34,7 +34,7 @@ namespace nutty
 
             uint rest = (d % elementsPerBlock);
 
-            uint grid = d / elementsPerBlock;
+            uint grid = (uint)d / elementsPerBlock;
 
             if(rest > 0)
             {
